+++ CHANGELOG.rst
@@ -7,6 +7,11 @@ Change log
 Next version
 ~~~~~~~~~~~~
 
+- ``ImportMap`` now copies the data it is given, so changing the original dict
+  no longer changes the import map (and its hash).
+- Deprecated ``ImportMap.update()``, import maps will become immutable. Use
+  ``map1 | map2`` or ``map1 |= map2`` instead.
+
 4.1 (2026-08-28)
 ~~~~~~~~~~~~~~~~